Handle more obscure tar buffer cases.
In some cases, (notably, when using file:// URLs) the existing code to skip past
the headers failed to work properly. By jumping to the beginning of the file
before skipping the headers, some of the edge cases seem to have been dealt
with. If point was not correctly set before calling `package-unpack', the tar
library would see a corrupted tar file and die.

Signed-off-by: Dan Hackney <dan@haxney.org>