package-test.el: Begin work on testing environment.
Uses `el-expectations', which looks to be powerful enough and well-supported.
The biggest thing to note at this stage is the `with-package-test' macro which
can be used to create a sort of testing fixture by let-binding values for
variables. This allows testing to be done without clobbering the current
session's variables.

Signed-off-by: Dan Hackney <dan@haxney.org>