Add function `package-from-dirname'.
Creates a minimal package structure from a directory by looking for an archive
which contains the directory. Also takes advantage of the fact that the package
name and version are included in the directory name.

This also changes `package-load-descriptor' to require a single PKG argument
which is a skeleton package structure containing (at a minimum) :name, :version,
and :archive.

Signed-off-by: Dan Hackney <dan@haxney.org>