package-install-from-buffer: Better handling of existing package data.
Rather than refusing to parse the buffer when given an existing package,
overwrite the given package with those values found in the buffer. In some
cases, not all of the package data can be extracted from the buffer (such as the
package name), so the supplied package structure might not be completely
overwritten.

Signed-off-by: Dan Hackney <dan@haxney.org>