package-from-file: Pull additional info from `package-from-filename'.
Originally, `package-from-file' did not consider the name of the file when
reading information, only the contents of the file. Some useful information,
like the name of the package and its archive, are typically only available in
the file name. Try to pick up as much of this information as is available.

Signed-off-by: Dan Hackney <dan@haxney.org>