package-initialize: Register from archives first.
Register packages from the archives before registering installed or activated
packages. This ensures that packages will have the correct status, since all
packages in the archive will be listed as `available'. By registering the
archive contents first, the status of installed and activated packages is
overwritten with the correct value.

Signed-off-by: Dan Hackney <dan@haxney.org>